Add some type-annotations for JET

+++ src/ints.jl
@@ -35,7 +35,7 @@ overflowval(::Type{T}) where {T <: Integer} = div(typemax(T) - T(9), T(10))
         end
         if has_groupmark
             b, nb = dpeekbyte(source, pos) .- UInt8('0')
-            if options.groupmark - UInt8('0') == b && nb <= 0x09
+            if (options.groupmark)::UInt8 - UInt8('0') == b && nb <= 0x09
                 incr!(source)
                 pos += 1
                 b = nb
@@ -78,7 +78,7 @@ overflowval(::Type{T}) where {T <: Integer} = div(typemax(T) - T(9), T(10))
         end
         if has_groupmark
             b, nb = dpeekbyte(source, pos) .- UInt8('0')
-            if options.groupmark - UInt8('0') == b && nb <= 0x09
+            if (options.groupmark)::UInt8 - UInt8('0') == b && nb <= 0x09
                 incr!(source)
                 pos += 1
                 b = nb
